+++ stove/templates/content_tracker_base.html
@@ -4,7 +4,6 @@
 
 {% block stylesheet %}
 <link rel="stylesheet" href="{% static 'ubyssey/css/stove/content_tracker.css' %}"></link>
-<script defer src="{% static 'ubyssey/js/stove/content_tracker.js' %}"></script>
 
 
 {% endblock %}
