+++ ubyssey/static_src/src/stove/StoveContentTracker.jsx
@@ -1,11 +1,6 @@
 import { useState, useEffect } from "react";
 import { createRoot } from 'react-dom/client';
-import { flushSync } from 'react-dom';
-import Select from 'react-select';
 
-import DatePicker from 'react-datepicker';
-import "react-datepicker/dist/react-datepicker.css";
-import chroma from 'chroma-js';
 import Switch from "react-switch";
 
 import { Group, Panel, Separator} from "react-resizable-panels";
@@ -16,891 +11,28 @@ import Tab from 'react-bootstrap/Tab';
 import Tabs from 'react-bootstrap/Tabs';
 import Table from 'react-bootstrap/Table';
 import Button from 'react-bootstrap/Button';
-import SvgStoveNameplateBlue from './stove-nameplate-blue.svg';
 
 import Skeleton from 'react-loading-skeleton'
 import 'react-loading-skeleton/dist/skeleton.css'
 
-import "prosemirror-view/style/prosemirror.css";
-import { EditorState } from "prosemirror-state";
-import { history } from 'prosemirror-history'
-import { Schema } from 'prosemirror-model'
-import { toggleMark, joinBackward } from 'prosemirror-commands'
-import { undo, redo } from 'prosemirror-history'
-import { baseKeymap } from 'prosemirror-commands'
-import { keymap } from 'prosemirror-keymap'
-import { marks } from 'prosemirror-schema-basic'
-import { DOMParser } from "prosemirror-model";
-import { autolink } from "prosemirror-autolink";
-import { Tooltip } from 'react-tooltip';
-
-
-
-
-import {
-  ProseMirror,
-  ProseMirrorDoc,
-  reactKeys,
-} from "@handlewithcare/react-prosemirror";
-
+import Deadline, {getDeadlineDate} from "./assignment_manager/Deadline.jsx";
+import DateInput from "./assignment_manager/DateInput.jsx";
+import deadlineOptions from './assignment_manager/DeadlineOptions.json'
+import EthicsNotes from "./assignment_manager/EthicsNotes.jsx";
+import AssignmentMemo from "./assignment_manager/AssignmentMemo.jsx";
+import SectionSelect from "./assignment_manager/SectionSelect.jsx";
+import {updateArticleStatus, updateAuthors, updateBeat, updateStoryType, updateSection, updateDeadline, updateTitle, remoteCreatePage} from "./assignment_manager/remoteManagement.js";
+import StoryTypeSelect, {storyTypeLabel} from "./assignment_manager/StoryTypeSelect.jsx";
+import BeatSelect from "./assignment_manager/BeatSelect.jsx";
+import LinkInput, {LinkOpenButton} from "./assignment_manager/LinkInput.jsx";
+import AuthorsSelect from "./assignment_manager/AuthorSelect.jsx";
+import ArticleStatus from "./assignment_manager/ArticleStatus.jsx";
+import statuses from './assignment_manager/statuses.json'
+import NavigationSidebar from "./assignment_manager/NavigationSidebar.jsx";
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import { ToastContainer, toast } from "react-toastify";
-// import { newDate } from "react-datepicker/dist/dist/date_utils.js";
-
-
-const groupStyles = {
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-};
-const groupBadgeStyles = {
-  backgroundColor: '#EBECF0',
-  borderRadius: '2em',
-  color: '#172B4D',
-  display: 'inline-block',
-  fontSize: 12,
-  fontWeight: 'normal',
-  lineHeight: '1',
-  minWidth: 1,
-  padding: '0.16666666666667em 0.5em',
-  textAlign: 'center',
-};
-const chosenColor = '#FFD230'
-
-const colourStyles = {
-    control: (styles) => ({ ...styles, backgroundColor: chosenColor }),
-    option: (styles, { data, isDisabled, isFocused, isSelected }) => {
-    const chromaColor = chroma('#FFD230');
-    return {
-      ...styles,
-      backgroundColor: isDisabled
-        ? undefined
-        : isSelected
-        ? chosenColor
-        : isFocused
-        ? chromaColor.alpha(0.1).css()
-        : undefined,
-      color: isDisabled
-        ? '#ccc'
-        : isSelected
-        ? chroma.contrast(chromaColor, 'white') > 2
-          ? 'white'
-          : 'black'
-        : chosenColor,
-      cursor: isDisabled ? 'not-allowed' : 'default',
-
-      ':active': {
-        ...styles[':active'],
-        backgroundColor: !isDisabled
-          ? isSelected
-            ? chosenColor
-            : chromaColor.alpha(0.3).css()
-          : undefined,
-      },
-    };
-  },
-};
-
-function findAuthorName(authorId) {
-  for (let entry in authors) {
-    const author = authors[entry]
-    if (author["value"] == authorId) {
-      return author["label"]
-    }
-  }
-  return "";
-}
-  const responsibleRole = 
-    [
-      {role: "author", color: "#e6e6e6"},
-      {role: "author", color: "#e6e6e6"},
-      {role: "backfield_editor", color: "#f5c554"},
-      {role: "author", color: "#e6e6e6"},
-      {role: "copy_editor", color: "#77c0d2"},
-      {role: "author", color: "#e6e6e6"},
-      {role: "author", color: "#77d27cff"}
-    ];
-  const roleColours = {
-    "author": "#e6e6e6",
-    "backfield_editor": "#f5c554",
-    "copy_editor": "#77c0d2",
-    "published_author":  "#c0e5bd"
-  }
-
-function AuthorsSelect ({currentAuthors, handleUpdateAuthors, authorType, styleType="edit-field", disabled, isPublished}) {
-  if (currentAuthors === "") {
-    return <Skeleton width="14em"/>
-  }
-  
-  let initialAuthors = [];
-  for (const authorId in currentAuthors) {
-    const author = currentAuthors[authorId]
-    if (author["author_role"] == authorType) {
-      initialAuthors.push({value: author["author"], label: findAuthorName(author["author"])})
-    }
-  }
-
-  let style = {
-    multiValue: (base) => ({
-        ...base,
-        backgroundColor: isPublished ? roleColours["published_author"] : roleColours[authorType],
-    }),
-    menu: (base) => ({
-      ...base,
-      marginTop: "-4px"
-    })};
-
-  if (styleType == "edit-field") {
-    style = {
-      ...style, 
-      control: (base) => ({
-        ...base,
-        border: "none",
-        backgroundColor: "inherit",
-      }),
-      valueContainer: (base) => ({
-        ...base,
-        padding: "5px",
-        ':hover': {
-          backgroundColor: "var(--hover-color)"
-        }
-      }),
-      selectContainer: (base) => ({
-        ...base,
-        padding: "0",
-        margin: "0",
-      })
-      
-    }
-  } else {
-    style = {
-      ...style, 
-      container: (base) => ({
-        ...base,
-        maxWidth: "20em",
-      })
-    }
-  }
-  
-  if (disabled) {
-    style = {
-      ...style,
-      container: (base) => ({
-        ...base,
-        pointerEvents: "auto",
-      }),
-      valueContainer: (base) => ({
-        ... base,
-        ':hover': {
-          cursor: "not-allowed",
-          backgroundColor: "var(--invalid-hover-color)"
-        },
-        ':active': {
-          pointerEvents: "none",
-          backgroundColor: "var(--invalid-hover-color)"
-        }
-      }),
-      multiValueRemove: (base) => ({
-        ...base,
-        ':hover': {
-          cursor: "not-allowed",
-          backgroundColor: "inherit"
-        },
-      })
-    }
-  }
-
-  return <Select 
-    isDisabled = {disabled}
-    options={authors} 
-    onChange = {handleUpdateAuthors} 
-    value={initialAuthors} 
-    isMulti 
-    styles={style}
-    placeholder = {"Add " + authorType.replace("_", " ") + "..."}
-    components={{
-      DropdownIndicator: null, 
-      ClearIndicator: null
-    }}/>
-}
-
-function getRandomInt(max) {
-  return Math.floor(Math.random() * max);
-}
-
-function resolveAfter2Seconds() {
-  return new Promise((resolve, reject) => {
-    setTimeout(() => {
-      if (getRandomInt(10)) {
-        resolve("resolved");
-      } else {
-        reject("failed")
-      }
-    }, getRandomInt(5000));
-  });
-}
-
-async function updateTitle(page, newTitle, updatePage) {
-
-  page.title = newTitle
-
-  updatePage(page)
-
-  handleRemoteUpdate(page, {"title": page.title}, updatePage,
-    "Updating title for " + page.title, 
-    "Updated title for " + page.title, 
-    "Failed to title for " + page.title)
-
-}
-
-function isValidUrl(url) {
-  const urlRegex = /https?:\/\/..*\...*/g
-
-  return url.match(urlRegex)
-}
-
-async function updateAssignmentFolder(page, newAssignmentFolder, updatePage, isLocalOnly=false) {
-  if (page.assignment_folder == newAssignmentFolder) return;
-
-  if (newAssignmentFolder.indexOf("http://") != 0 && newAssignmentFolder.indexOf("https://") !=0) {
-    newAssignmentFolder = "http://" + newAssignmentFolder;
-  }
-  if (!isValidUrl(newAssignmentFolder)) {
-    toast.error(
-      newAssignmentFolder + " is not a valid URL.", 
-      {position: 'bottom-left'}
-    )
-    return;
-  }
-
-  page.assignment_folder = newAssignmentFolder
-
-  updatePage(page)
-
-  if (!isLocalOnly) {
-    handleRemoteUpdate(page, {"assignment_folder": newAssignmentFolder}, updatePage,
-      "Updating assignment folder for " + page.title, 
-      "Updated assignment folder for " + page.title, 
-      "Failed to update assignment folder for " + page.title)
-  }
-}
-
-
-async function updateAuthors(page, newAuthorList, role, updatePage, isLocalOnly=false) {
-
-  page.article_authors = page.article_authors.filter((author) => author.author_role != role)
-
-  for (const author of newAuthorList) {
-    page.article_authors.push({article_page: page.pk, author_role: role, author: author.value})
-  } 
-
-  updatePage(page)
-
-  if (!isLocalOnly) {
-    handleRemoteUpdate(page, {"authors": page.article_authors}, updatePage,
-      "Updating authors for " + page.title, 
-      "Updated authors for " + page.title, 
-      "Failed to authors for " + page.title)
-  }
-
-}
-
-
-async function updateArticleStatus(page, newStatus, updatePage, isLocalOnly=false) {
-  updatePage({... page, article_status: newStatus.value})
-  
-  if (!isLocalOnly) {
-    handleRemoteUpdate(page, {"article_status": newStatus.value}, updatePage,
-      "Updating status for " + page.title + " to " + newStatus.label, 
-      "Updated status for " + page.title + " to " + newStatus.label, 
-      "Failed to update status for " + page.title + " to " + newStatus.label)
-  }
-}
-
-function getDeadlineDate(page, deadlineDescription) {
-  for (const deadline of page.deadline_list) {
-    if (deadline.description == deadlineDescription) {
-      return deadline.date;
-    }
-  }
-  return null;
-}
-
-function getDeadlineByDescription(page, deadlineDescription) {
-  for (const deadline of page.deadline_list) {
-    if (deadline.description == deadlineDescription) {
-      return deadline;
-    }
-  }
-  return null;
-}
-
-async function updateDeadlineList(page, newDeadlineList, updatePage, isLocalOnly=false) {
-  updatePage({... page, deadline_list: newDeadlineList})
-
-  if (!isLocalOnly) {
-    handleRemoteUpdate(page, {"deadline_list": newDeadlineList}, updatePage,
-      "Updating deadline for " + page.title, 
-      "Updated deadline for " + page.title, 
-      "Failed to update deadline for " + page.title)
-  }
-}
-
-const DEADLINE = {
-  DRAFT_IN: "Draft in",
-  RESEARCH: "Research",
-  QUESTIONS: "Questions",
-  SOURCES: "Sources",
-  EDITING: "Editing",
-  PUBLISHING: "Publishing"
-}
-
-async function updateDeadline(page, newDeadline, updatePage, isLocalOnly=false) { //TODO
-  let {date, description, completed} = newDeadline
-  let deadlineList = [...page.deadline_list]
-  let updated = false
-
-  if (typeof date !== "string") {
-    date=date.toISOString()
-  }
-
-  deadlineList = deadlineList.map(deadline => {
-    if (deadline.description !== description) {
-      return deadline
-    } else {
-      updated=true;
-      deadline.date = date;
-      deadline.completed = completed;
-      return deadline;
-    }
-  })
-  if (!updated) {
-    deadlineList = [... deadlineList,
-      {
-        completed: completed,
-        date: date,
-        description: description
-      }
-    ]
-  }
-
-  updateDeadlineList(page, deadlineList, updatePage, isLocalOnly)
-
-}
-
-async function updateBeat(page, newBeat, updatePage, isLocalOnly=false) {
-  updatePage({... page, category_page: newBeat.value})
-
-  if (!isLocalOnly) {
-  handleRemoteUpdate(page, {"category": newBeat.label}, updatePage,
-    "Updating beat for " + page.title + " to " + newBeat.label, 
-    "Updated beat for " + page.title + " to " + newBeat.label, 
-    "Failed to update beat for " + page.title + " to " + newBeat.label)
-  }
-}
-async function updateStoryType(page, newStoryType, updatePage, isLocalOnly=false) {
-  updatePage({... page, story_type: newStoryType.value})
-
-  if (!isLocalOnly) {
-  handleRemoteUpdate(page, {"story_type": newStoryType.value}, updatePage,
-    "Updating story type for " + page.title + " to " + newStoryType.label, 
-    "Updated story type for " + page.title + " to " + newStoryType.label, 
-    "Failed to update story type for " + page.title + " to " + newStoryType.label)
-  }
-}
-
-async function updateSection(page, newSection, updatePage, isLocalOnly=false) {
-  updatePage({... page, current_section: newSection.value})
-
-  if (!isLocalOnly) {
-    handleRemoteUpdate(page, {"current_section": newSection.value}, updatePage,
-      "Updating beat for " + page.title + " to " + newSection.label, 
-      "Updated beat for " + page.title + " to " + newSection.label, 
-      "Failed to update beat for " + page.title + " to " + newSection.label)
-  }
-}
-
-async function updateAssignmentMemo(page, newMemo, updatePage, isLocalOnly = false) {
-  if (page["assignment_memo"] == newMemo) return;
-
-  updatePage({...page, assignment_memo: newMemo})
-
-  if(!isLocalOnly) {
-    handleRemoteUpdate(page, {"assignment_memo": newMemo}, updatePage,
-      "Updating assignment memo for " + page.title, 
-      "Updated assignment memo for " + page.title, 
-      "Failed to assignment memo for " + page.title)
-  }
-
-}
-
-async function updateEthicsNotes(page, newEthics, updatePage, isLocalOnly = false) {
-  if (page.ethics_notes == newEthics) return;
-
-  updatePage({...page, ethics_notes: newEthics})
-
-
-  if(!isLocalOnly) {
-    handleRemoteUpdate(page, {"ethics_notes": newEthics}, updatePage,
-      "Updating ethics notes for " + page.title, 
-      "Updated ethics notes for " + page.title, 
-      "Failed to update ethics notes for " + page.title)
-  }
-}
-
-async function handleRemoteUpdate(page, changes, updatePage, pendingText, successText, errorText) {
-  toast.promise(
-      remoteUpdatePage(page, changes),
-      {
-        pending: {
-          render(){
-            return pendingText
-          },
-          position: 'bottom-left',
-        },
-        success: {
-          render() {return successText},
-          autoClose: 1500},
-        error: errorText,
-        
-      }
-    )
-    .then((result) => {
-      updatePage(JSON.parse(result))
-    }
-    )
-    .catch((error) => console.log(error))
-}
-
-async function remoteUpdatePage(page, body) {
-  let headers = {content_type: "application/json"}
-
-  for (const key of Object.keys(requiredHeader)) {
-      headers[key] = requiredHeader[key]
-  }
-
-  return fetch(updateEndpoint.replace("1918", page.pk), { method: "POST", headers: headers, body: JSON.stringify(body),
-    credentials: "same-origin"})
-    .then(async (response) => {
-      if (response.status != 200) {
-        throw new Error(response)
-      } else {
-        return response.json()
-      }}
-    );
-  // const payload = await response.json();
-}
-
-async function remoteCreatePage(page) {
-  let headers = {content_type: "application/json"}
-
-  for (const key of Object.keys(requiredHeader)) {
-      headers[key] = requiredHeader[key]
-  }
-
-  if (!page.current_section) {
-    throw new Error("Pages must have a section")
-  }
-  return fetch(createPageUrl.replace("1918", page.current_section), { method: "POST", headers: headers, body: JSON.stringify(page),
-    credentials: "same-origin"})
-    .then(async (response) => {
-      if (response.status != 200) {
-        const errorText = await response.text()
-        throw new Error(errorText)
-      } else {
-        return response.json()
-      }}
-    );
-  // const payload = await response.json();
-}
-
-const formatGroupLabel = (data) => (
-  <div style={groupStyles}>
-    <span>{data.label}</span>
-    <span style={groupBadgeStyles}>{data.options.length}</span>
-  </div>
-);
-
-const articleStatusStyles = {
-  control: (styles) => ({ ...styles, backgroundColor: 'white' }),
-  option: (styles, { data, isDisabled, isFocused, isSelected }) => {
-    const color = chroma(data.color);
-    return {
-      ...styles,
-      backgroundColor: isDisabled
-        ? undefined
-        : isSelected
-        ? data.color
-        : isFocused
-        ? color.alpha(0.1).css()
-        : undefined,
-      color: isDisabled
-        ? '#ccc'
-        : isSelected
-        ? chroma.contrast(color, 'white') > 2
-          ? 'white'
-          : 'black'
-        : data.color,
-      cursor: isDisabled ? 'not-allowed' : 'default',
-
-      ':active': {
-        ...styles[':active'],
-        backgroundColor: !isDisabled
-          ? isSelected
-            ? data.color
-            : color.alpha(0.3).css()
-          : undefined,
-      },
-    };
-  }
-}
-const statuses = 
-    [
-      {value: 1, label: "Assigned", color: '#ea968d', textColor: "#691b17"},
-      {value: 2, label: "Filed", color: '#f9de9d', textColor: "#80511f"},
-      {value: 3, label: "Edited", color: '#f5c554', textColor: "#80511f"},
-      {value: 4, label: "Copy", color: '#77c0d2', textColor: "#172448"},
-      {value: 5, label: "Ready", color: '#e3e455', textColor: '#264b35'},
-      {value: 6, label: "Published", color: '#9ec756', textColor: '#264b35'}
-    ];
-
-function ArticleStatus ({status, updateStatus}) {
-
-  
-  return <Select 
-    className="status-select"
-    options={statuses.slice(0, -1)} 
-    value={statuses[status-1]} 
-    styles={{
-      singleValue: (base) => ({
-        ...base,
-        padding: 5,
-        borderRadius: 5,
-        background: statuses[status-1].color,
-        color: statuses[status-1].textColor,
-        fontWeight: "bold",
-        textAlign: "center",
-      }),
-      valueContainer: (base) => ({
-        ...base,
-        padding: 0,
-      }),
-      control: (base) => ({
-        ...base,
-        border: "none",
-        background: "none",
-        boxShadow: "none",
-      }),
-      menu: (base) => ({
-        ...base,
-        marginTop: "-4px"
-      })
-    }}
-    onChange={updateStatus}
-    isDisabled= {status==6}
-    components={{
-      DropdownIndicator: null, 
-      placeholder: "Select status..."}}  
-  />;
-}
-
-function beatLabel(beatPk) {
-
-  for (const {label, options} of beatOptions) {
-    for (const beat of options) {
-      if (+beatPk === beat.value) return beat.label
-    }
-  }
-  return "[No label provided]"
-}
-
-const storyTypeOptions = [
-  {value: 'report', label: "Report", description: "Reports are shorter stories about events with immediate relevance, written from a detached perspective."},
-  {value: 'live-update', label: "Live Update", description: "Live Updates are brief reports from journalists on-the-ground while news is happening."},
-  {value: 'feature', label: "Feature", description: "Features are longer stories about people or systems with long-term or widespread relevance, written from a reporter's perspective."},
-  {value: 'profile', label: "Profile", description: "Profiles tell the stories of individuals and their worldviews, written from a repoter's perspective."},
-  {value: 'qa', label: "Q&A", description: "Q&As are a transcription of a conversation between an interviewee and The Ubyssey, edited by our journalists for length and clarity."},
-  {value: 'review', label: "Review", description: "Reviews are storiess about art and culture, written from a critical perspective."},
-  {value: 'game-analysis', label: "Game Analysis", description: "Game analyses are stories about individual games, written from an reporter's perspective."},
-  {value: 'commentary', label: "Commentary", description: "[Description TK]"},
-  {value: 'essay', label: "Essay", description: "Essays are stories about the author's views on the news, written from their own perspective but based on reporting."},
-  {value: 'column', label: "Column", description: "Columns are stories reported by columnists, opinion journalists who make abstract judgments about the news and the way the world should be."},
-  {value: 'editorial', label: "Editorial", description: "Editorials are stories by the Editorial Board, the body of the newspaper's staff who debate and decide positions on the news of the day."},
-  {value: 'letter-to-editor', label: "Letter to the Editor", description: "Letters to the editor are 250-word responses to stories published in The Ubyssey, written by readers."},
-  {value: 'letter-from-editor', label: "Letter from the Editor", description: "Letters from the editor messages to readers from the The Ubyssey's Senior Masthead."},
-  {value: 'psa', label: "Public Service Announcement", description: "Annoucements of public interest information with imminent relevance, such as extreme weather or threats to public safety."},
-  {value: 'other', label: "Other", description: "Editors must seek approval from the senior masthead before publishing an 'other' story type."},
-    ]
-
-  function storyTypeLabel(storyTypeValue) {
-    for (const {value, label} of storyTypeOptions) {
-      if (value === storyTypeValue) return label;
-    }
-    return "[No label provided]"
-  }
-
-function StoryTypeSelect ({storyType, updateStoryType, styleType="edit-field", disabled}) {
-
- let style = {
-  ...style,
-  menu: (base) => ({
-    ...base,
-    marginTop: "-4px"
-  })
- };
-
-  if (styleType == "edit-field") {
-    style = {
-      ...style, 
-      control: (base) => ({
-        ...base,
-        border: "none",
-        backgroundColor: "inherit",
-      }),
-      valueContainer: (base) => ({
-        ...base,
-        padding: "5px",
-        ':hover': {
-          backgroundColor: "var(--hover-color)"
-        }
-      }),
-      selectContainer: (base) => ({
-        ...base,
-        padding: "0",
-        margin: "0",
-      }), 
-      container: (base) => ({
-        ...base,
-        maxWidth: "20em",
-      })
-    }
-  }
-
-  if (disabled) {
-    style = {
-      ...style,
-      container: (base) => ({
-        ...base,
-        pointerEvents: "auto",
-      }),
-      valueContainer: (base) => ({
-        ... base,
-        ':hover': {
-          cursor: "not-allowed",
-          backgroundColor: "var(--invalid-hover-color)"
-        },
-        ':active': {
-          pointerEvents: "none",
-          backgroundColor: "var(--invalid-hover-color)"
-        }
-      }),
-      singleValue: (base) => ({
-        ...base,
-        color: "inherit"
-      })
-    }
-  }
-
-  const formatOptionLabel = ({ value, label, description }) => (
-    <div style={{ display: "grid", gridTemplateColumns: "auto" }}>
-      <div>{label}</div>
-      <div style={{ color: "#555555", fontSize: "small" }}>
-        <i>{description}</i>
-      </div>
-    </div>
-  );
-
-    return <Select 
-    isDisabled={disabled}
-    options={storyTypeOptions}
-    value={storyType ? {"value": storyType, "label": storyTypeLabel(storyType)} : ''}
-    onChange={updateStoryType}
-    styles={style}
-    formatGroupLabel={formatGroupLabel}
-    formatOptionLabel={formatOptionLabel}
-    components={{
-      DropdownIndicator: null, 
-      placeholder: "Choose story type..."}}/>
-}
-function BeatSelect ({beat, updateBeat, styleType="edit-field", disabled}) {
-
- let style = {
-  ...style,
-  menu: (base) => ({
-    ...base,
-    marginTop: "-4px"
-  })
- };
-
-  if (styleType == "edit-field") {
-    style = {
-      ...style, 
-      control: (base) => ({
-        ...base,
-        border: "none",
-        backgroundColor: "inherit",
-      }),
-      valueContainer: (base) => ({
-        ...base,
-        padding: "5px",
-        ':hover': {
-          backgroundColor: "var(--hover-color)"
-        }
-      }),
-      selectContainer: (base) => ({
-        ...base,
-        padding: "0",
-        margin: "0",
-      }), 
-      container: (base) => ({
-        ...base,
-        maxWidth: "20em",
-      })
-    }
-  }
-
-  if (disabled) {
-    style = {
-      ...style,
-      container: (base) => ({
-        ...base,
-        pointerEvents: "auto",
-      }),
-      valueContainer: (base) => ({
-        ... base,
-        ':hover': {
-          cursor: "not-allowed",
-          backgroundColor: "var(--invalid-hover-color)"
-        },
-        ':active': {
-          pointerEvents: "none",
-          backgroundColor: "var(--invalid-hover-color)"
-        }
-      }),
-      singleValue: (base) => ({
-        ...base,
-        color: "inherit"
-      })
-    }
-  }
-
-    return <Select 
-    isDisabled={disabled}
-    options={beatOptions}
-    value={beat ? {"value": beat, "label": beatLabel(beat)} : ''}
-    onChange={updateBeat}
-    styles={style}
-    formatGroupLabel={formatGroupLabel}
-    components={{
-      DropdownIndicator: null, 
-      placeholder: "Choose beat..."} }/>
-}
-function findSection(sectionSlug) {
-    for (const s of allSections) {
-      if (s.slug == sectionSlug) return s
-    }
-    return undefined
-  }
-
-function SectionSelect ({section, updateSection, styleType="edit-field"}) {
-
-  let style = {};
-
-  if (styleType == "edit-field") {
-    style = {
-      ...style, 
-      control: (base) => ({
-        ...base,
-        border: "none",
-        backgroundColor: "inherit",
-      }),
-      valueContainer: (base) => ({
-        ...base,
-        padding: "5px",
-        ':hover': {
-          backgroundColor: "var(--hover-color)"
-        }
-      }),
-      selectContainer: (base) => ({
-        ...base,
-        padding: "0",
-        margin: "0",
-      }), 
-      container: (base) => ({
-        ...base,
-        maxWidth: "20em",
-      })
-    }
-  }
-
-  return <Select 
-    options={allSections}
-    value={ section ? findSection(section) : ''}
-    onChange={updateSection}
-    styles={style}
-    formatGroupLabel={formatGroupLabel}
-    components={{
-      DropdownIndicator: null, 
-      placeholder: "Choose section..."} }/>
-}
-
-function LinkInput ({selectedPage, updatePage, isLocalOnly=false}) {
-  const [assignmentFolder, changeAssignmentFolder] = useState(selectedPage.assignment_folder);
-
-  useEffect(() => {
-    if (selectedPage.assignment_folder != null) {
-      changeAssignmentFolder(selectedPage.assignment_folder)
-    } else {
-      changeAssignmentFolder('')
-    }
-  }, [selectedPage]);
-
-  return <div class="edit-field--hyperlink-container">
-          <LinkOpenButton url={assignmentFolder} className={"edit-field--hyperlink-open"}/>
-          <input class="edit-field--plaintext" 
-            placeholder="Assignment folder link..." 
-            value={assignmentFolder}
-            onChange={e => changeAssignmentFolder(e.target.value)}
-            onBlur={(e) => updateAssignmentFolder(selectedPage, e.target.value, updatePage, isLocalOnly)}>
-            </input>
-          </div>
-}
-
-function LinkOpenButton ({url, className, iconSize}) {
-  let button;
-
-  if (url != null && url != '') {
-    const activeButton = isValidUrl(url)
-
-    if (isValidUrl(url) || isValidUrl("http://" + url)) {
-      button = <a href={url} target="_blank" rel="noopener noreferrer">
-          <FolderOpen 
-            width={iconSize}
-            height={iconSize}
-          />
-        </a>
-    } else {
-      button = <FolderOpenOutline 
-        color={"#d23732"}
-        width={iconSize}
-        height={iconSize}/>
-    }
-  } else {
-    button = <FolderOpenOutline 
-        width={iconSize}
-        height={iconSize}/>
-  }
-
-  return (<div className={className}>
-             {button}
-            </div>);
-}
-
-
-
-
 
 function ArticleRow({page, updatePage, selectedArticleId, setSelectedArticleId, setActiveSidebar}) {
 
@@ -917,6 +49,7 @@ function ArticleRow({page, updatePage, selectedArticleId, setSelectedArticleId, 
       setSelectedArticleId(page.pk)
       setActiveSidebar(mode)
     }
+
 
     return <tr key={page.pk} className={selectedClass}>
             <td class="slug-cell">
@@ -944,12 +77,12 @@ function ArticleRow({page, updatePage, selectedArticleId, setSelectedArticleId, 
             <td class="authors-cell"><AuthorsSelect 
               disabled = {page.live}
               currentAuthors={page.article_authors} 
-              handleUpdateAuthors={(newAuthorList) => updateAuthors(page, newAuthorList, responsibleRole[page.article_status].role, updatePage)}
-              authorType={responsibleRole[page.article_status].role}
+              handleUpdateAuthors={(newAuthorList) => updateAuthors(page, newAuthorList, statuses[page.article_status].role, updatePage)}
+              authorType={statuses[page.article_status].role}
               isPublished={page.live}
               />
             </td>
-            <td><DateInput date={getDeadlineDate(page, DEADLINE.DRAFT_IN)} handleUpdateDate={(newDate) => updateDeadline(page, {date: newDate, description: DEADLINE.DRAFT_IN, completed: false}, updatePage)} disabled={page.live}/></td>
+            <td><DateInput date={getDeadlineDate(page, deadlineOptions.DRAFT_IN)} handleUpdateDate={(newDate) => updateDeadline(page, {date: newDate, description: deadlineOptions.DRAFT_IN, completed: false}, updatePage)} disabled={page.live}/></td>
             <td><BeatSelect beat={page.category_page} updateBeat={(newBeat) => updateBeat(page, newBeat, updatePage)} disabled={page.live}/></td>
             <td><ArticleStatus status={page["article_status"]} updateStatus={(newStatus) => updateArticleStatus(page, newStatus, updatePage)}/></td>
             <td>
@@ -966,7 +99,6 @@ function ArticleRow({page, updatePage, selectedArticleId, setSelectedArticleId, 
 
 
 function ArticleList({allPages, updatePage, selectedArticleId, setSelectedArticleId, setActiveSidebar}) {
-
 
   const rows = []
   
@@ -1294,7 +426,7 @@ function CreateSidebar({createPage}) {
           <LinkInput selectedPage={newPage} updatePage={(e) => updateNewPage(e)} isLocalOnly={true}/>
         <div className="edit-field--side-label">Status</div> <ArticleStatus status={newPage["article_status"]} updateStatus={(newStatus) => updateArticleStatus(newPage, newStatus, (e) => updateNewPage(e), true)}/>
       </div>
-        <DeadlineListEdit
+        <Deadline
           page={newPage}
           updatePage={updateNewPage}
           isLocalOnly={true}
@@ -1371,258 +503,13 @@ function EditSidebar({selectedPage, updatePage}) {
           <LinkInput selectedPage={selectedPage} updatePage={updatePage}/>
         <div className="edit-field--side-label">Status</div> <ArticleStatus status={selectedPage["article_status"]} updateStatus={(newStatus) => updateArticleStatus(selectedPage, newStatus, updatePage)}/>
       </div>
-      <DeadlineListEdit page={selectedPage} updatePage={updatePage}/>
+      <Deadline page={selectedPage} updatePage={updatePage}/>
       <h5>Assignment Notes </h5>
       <AssignmentMemo selectedPage={selectedPage} updatePage={updatePage}/>
       <h5>Ethics Notes </h5>
       <EthicsNotes selectedPage={selectedPage} updatePage={updatePage}/>
     </div>
   </div>
-}
-
-function DeadlineCheckbox({completed, updateChecked, invalid = false}) {
-  return <div className={`edit-field--checkbox ${invalid ? "edit-field--checkbox-invalid" : ""}`}
-    onClick={()=>updateChecked(!completed)}>
-    {completed ? <CheckboxOutline height="1lh"/> : <SquareOutline height={"1lh"}/>}
-  </div>
-}
-
-function DeadlineClear({clearDeadline}) {
-
-  return <div className={`edit-field--date-clear`}
-    onClick={clearDeadline}>
-      <CloseOutline height="1lh" />
-  </div>
-}
-
-function DeadlineItem({deadline, updateDeadline, clearDeadline, canEditName=false, overrideDescription=false}) {
-  let description, completed, date;
-  if (deadline != null) {
-    description = deadline.description;
-    completed = deadline.completed;
-    date = deadline.date;
-  } else {
-    description = ""
-    completed = false
-    date = null
-  }
-
-  return <div className={`edit-field--sidebyside ${completed ? "edit-field--checked" : ""}`}>
-    <div 
-      className="edit-field--side-label edit-field--deadline-entry edit-field--deadline-label"
-      >
-      <DeadlineCheckbox
-        completed = {completed}
-        updateChecked={(newValue) => updateDeadline({date: date, description: description, completed: newValue})}
-        invalid = {!date}
-        />
-      <div className={`edit-field--label-text ${canEditName ? "edit-field--label-editable" : ""}`} contentEditable={canEditName}>{overrideDescription ? "Deadline" : description.toString()}</div></div>
-    <div className={`edit-field--date ${date || canEditName ? "edit-field--date-clearable" : ""}`} >
-          <DateInput 
-            date={date} 
-            handleUpdateDate={(newDate) => updateDeadline({date: newDate, description: description, completed: completed})}
-            disabled={completed}/>
-          {date || canEditName ? <DeadlineClear 
-            clearDeadline={() => clearDeadline(deadline)}/> : ""}
-    </div>
-  </div>
-}
-
-function AddDeadlineItem({addDeadline}) {
-
-  function processInput(e) {
-    if (e.target.value) addDeadline({description: e.target.value, date: null, completed: false})
-    e.target.value = ""
-  }
-
-  return <div className={`edit-field--add-deadline`} >
-        <input 
-          className="edit-field--add-deadline-input" 
-          type="text" 
-          placeholder="Add..." 
-          onBlur={processInput}
-          onKeyDown={ (event) => {
-            if (event.key == "Enter") processInput(event)
-          }}></input></div>
-
-}
-
-function DeadlineListEdit({page, updatePage, isLocalOnly=false}) {
-  const [expanded, setExpanded] = useState(false);
-
-  useEffect(() => setExpanded(false), [page.pk])
-
-  function clearDeadline(deadline) {
-    let newDeadlineList = []
-    for (const existingDeadline of page.deadline_list) {
-      if (existingDeadline.description != deadline.description) newDeadlineList.push(existingDeadline)
-    }
-    updateDeadlineList(page, newDeadlineList, updatePage, isLocalOnly)
-  }
-
-  if (page.deadline_list.length <= 1 && !expanded) {
-    return <div className="edit-field--deadlines">
-        <DeadlineItem 
-          deadline={getDeadlineByDescription(page, DEADLINE.DRAFT_IN)}
-          updateDeadline={(newDeadline) => updateDeadline(page, {...newDeadline, description: DEADLINE.DRAFT_IN}, updatePage, isLocalOnly)}
-          overrideDescription={true}
-          clearDeadline={clearDeadline}
-        />
-      <div className="edit-field--more-deadlines" onClick={() => setExpanded(true)}>More...</div></div>
-  }
-  const deadlineHtml = []
-
-  if (expanded) {
-    let standardDeadlines = {}
-    for (const [key, label] of Object.entries(DEADLINE)) {
-      standardDeadlines[key] = getDeadlineByDescription(page, label)
-    }
-    for (const [key, deadline] of Object.entries(standardDeadlines)) {
-      deadlineHtml.push(
-        <DeadlineItem 
-          deadline={deadline ? deadline : {description: DEADLINE[key], completed: false, date: null}}
-          updateDeadline={(newDeadline) => updateDeadline(page, newDeadline, updatePage, isLocalOnly)}
-          overrideDescription={false}
-          clearDeadline={clearDeadline}
-        />
-      )
-    }
-  }
-
-  const deadlinesToDisplay = expanded ?
-    page.deadline_list.filter(({description}) => !Object.values(DEADLINE).includes(description)) :
-    page.deadline_list
-  for (const deadline of deadlinesToDisplay) {
-    deadlineHtml.push(
-      <DeadlineItem 
-          deadline={deadline}
-          updateDeadline={(newDeadline) => updateDeadline(page, newDeadline, updatePage, isLocalOnly)}
-          overrideDescription={false}
-          canEditName={!Object.values(DEADLINE).includes(deadline.description)}
-          clearDeadline={clearDeadline}
-        />)
-  }
-  return <>
-    <h5>Deadlines</h5>
-    <div className="edit-field--deadlines">
-
-        {deadlineHtml} 
-
-      {!expanded ? 
-        <div className="edit-field--more-deadlines" onClick={() => setExpanded(true)}>More...</div> : 
-        <AddDeadlineItem addDeadline={(newDeadline) => updatePage({...page, deadline_list: [...page.deadline_list, newDeadline]})}/>}
-    </div>
-    </>;
-
-}
-
-
-const doc = {
-  content: 'block+',
-  toDOM: () => ['article', 0],
-}
-
-const text = {
-  group: 'inline',
-}
-
-const paragraph = {
-  content: 'inline*',
-  group: 'block',
-  parseDOM: [{ tag: 'p' }],
-  toDOM: () => ['p', 0],
-}
-
-const schema = new Schema({
-  nodes: { doc, text, paragraph },
-  marks: marks,
-})
-
-function EthicsNotes({selectedPage, updatePage, isLocalOnly = false}) {
-  return <RichTextEditor onBlurCallback={(e) => updateEthicsNotes(selectedPage, e.target.innerHTML, updatePage, isLocalOnly)} defaultText={selectedPage.ethics_notes}/>
-}
-
-function AssignmentMemo({selectedPage, updatePage, isLocalOnly = false}) {
-  return <RichTextEditor onBlurCallback={(e) => updateAssignmentMemo(selectedPage, e.target.innerHTML, updatePage, isLocalOnly)} defaultText={selectedPage.assignment_memo}/>
-}
-
-function RichTextEditor({onBlurCallback, defaultText}) {
-  const domElement = new window.DOMParser().parseFromString(defaultText, "text/html").body;
-  const defaultNode = DOMParser.fromSchema(schema).parse(domElement);
-
-  const [editorState, setEditorState] = useState(
-    EditorState.create({
-        doc: defaultNode,
-        schema,
-        plugins: [
-          // The reactKeys plugin is required for the ProseMirror component to work!
-          reactKeys(),
-          history(),
-          keymap({
-            ...baseKeymap,
-            'Mod-z': undo,
-            'Shift-Mod-z': redo,
-            Backspace: joinBackward,
-            'Mod-b': toggleMark(schema.marks.strong),
-            'Mod-i': toggleMark(schema.marks.em)
-          }),
-          ...autolink({
-            openOnClick: true,
-            enableEnterTrigger: true,
-            excludedTrailingChars: ['.', ',', '!', '?', ':', ';', ')', ']', '}']
-          })
-        ],
-      })
-  )
-
-  useEffect(() => {
-    const newState = editorState;
-    newState.doc=defaultNode;
-    setEditorState(newState);
-  }, [defaultText]);
-  
-
-    return (
-    <ProseMirror
-      state={editorState}
-      dispatchTransaction={(tr) => {
-        setEditorState((s) => s.apply(tr));
-      }}
-    >
-      <div className="edit-field--richtext-editor">
-        
-      <ProseMirrorDoc 
-        onBlur={onBlurCallback}
-        style={{
-          backgroundColor: "#fdfdfd",
-          padding: "10px",
-          paddingBottom: "0",
-          minHeight: "6lh"
-        }}
-        />
-        <div className="edit-field--richtext-help"><a data-tooltip-id="richtext-info">?</a></div>
-        <Tooltip
-          id={"richtext-info"}
-          place={'top-end'}
-          style={{ 
-            backgroundColor: "#f6f6f6", 
-            fontSize: "small", 
-            color: "black", 
-            width: "200px", 
-            filter: 'drop-shadow(0px 4px 8px rgba(0, 0, 0, 0.2))'
-          }}
-        >
-          <span><strong>Shortcuts</strong></span>
-          <ul>
-            <li><strong>Bold</strong>. Cmd/Ctrl + B</li>
-            <li><i>Italics</i>. Cmd/Ctrl + I</li>
-            <li><a href="ubyssey.ca">Link</a>. Paste a link from your clipboard using Cmd/Ctrl + V</li>
-          </ul>
-        </Tooltip>
-        </div>
-        
-    </ProseMirror>
-  );
 }
 
 function SidebarViewsSelector({selectedPage, updatePage, createPage, activeSidebar, setActiveSidebar}) {
@@ -1671,22 +558,7 @@ function Sidebar({selectedPage, updatePage, createPage, activeSidebar, setActive
     </div>;
 }
 
-function SectionNavigationSidebar() {
-  function SectionGroup({groupName, sections}) {
-    const sectionItems = sections.map(section => <li className="section-navigation-item"><a href={"/stove/oven/" + section}>{section}</a></li>)
-    return <ul className="section-navigation-grouping">
-      <span className="section-navigation-title">{groupName}</span>
-      {sectionItems}
-    </ul>
-  }
-  return <div class="navigation-panel">
-            <div className="stove-logo-container"><a href="/stove/oven"><SvgStoveNameplateBlue className="stove-logo"/></a></div>
-          <SectionGroup groupName="Reportage" sections={["Arts", "Culture", "News", "Opinion", "Sports"]}/>
-          <SectionGroup groupName="Visuals" sections={["Graphics", "Photo", "Video"]}/>
-          <SectionGroup groupName="Product" sections={["Audio", "Print", "Socials"]}/>
-          <SectionGroup groupName="More" sections={["Copy", "Games", "Homepage"]}/>
-  </div>
-}
+
 
 function ContentTracker() {
   const [allPages, setAllPages] = useState(
@@ -1722,7 +594,7 @@ function ContentTracker() {
   return (
       <div className="content-tracker">
         <Group className="grouping">
-        <SectionNavigationSidebar/>
+        <NavigationSidebar/>
         <Panel className="panel main-content-panel" minSize="1100px" defaultSize="80%">
             <MainPanel 
               allPages={allPages} 
@@ -1754,27 +626,6 @@ function ContentTracker() {
   );
 }
 
-
-function DateInput ({date, handleUpdateDate, disabled}) {
-  return <div><DatePicker
-    disabled={disabled}
-    selected={date ? new Date(date) : undefined}
-    onChange={(newDate) => {
-      handleUpdateDate(newDate)
-    } }
-    showTimeSelect
-    timeFormat="h:mm aa"
-    timeIntervals={30}
-    timeCaption="time"
-    dateFormat="MMMM d, h:mm aa"
-    placeholderText="Add deadline"
-  /></div>;
-};
-
-
 const container = document.getElementById('content-tracker');
-const root = createRoot(container); // createRoot(container!) if you use TypeScript
+const root = createRoot(container); 
 root.render(<ContentTracker />);
-
-
-// https://codesandbox.io/p/sandbox/prosemirror-simple-i8yul?file=%2Fsrc%2Findex.ts%3A12%2C3
