+++ ubyssey/static_src/src/stove/assignment_manager/AssignmentMemo.jsx
@@ -1,0 +1,8 @@
+import RichTextEditor from "./RichTextEditor.jsx"
+import {updateAssignmentMemo} from "./remoteManagement.js";
+
+
+
+export default function AssignmentMemo({selectedPage, updatePage, isLocalOnly = false}) {
+  return <RichTextEditor onBlurCallback={(e) => updateAssignmentMemo(selectedPage, e.target.innerHTML, updatePage, isLocalOnly)} defaultText={selectedPage.assignment_memo}/>
+}
