+++ ubyssey/static_src/src/stove/assignment_manager/DeadlineOptions.json
@@ -1,0 +1,8 @@
+{
+  "DRAFT_IN": "Draft in",
+  "RESEARCH": "Research",
+  "QUESTIONS": "Questions",
+  "SOURCES": "Sources",
+  "EDITING": "Editing",
+  "PUBLISHING": "Publishing"
+}
