+++ ubyssey/static_src/webpack.common.js
@@ -39,9 +39,7 @@ module.exports = {
     'passing-2025': './src/js/passing-2025.js',
     'contentTracker': './src/js/ContentTracker.jsx',
     'liveblog': './src/js/liveblog.jsx',
-    'stove/content_tracker': './src/stove/content_tracker.js',
     'stove/manuscript_editor': './src/stove/manuscript/index.js',
-    'stove/content_tracker_select': './src/stove/ContentTrackerSelect.jsx',
     'stove/stove_content_tracker': './src/stove/StoveContentTracker.jsx',
     'join': './src/js/join.js',
   },
